Modify py_File so that it can accept binary as argument
The map script will open the .dat directly, and return the content as
base64 string. After conversion to binary, it will be fed to py_File.
So py_File is modified to be able to accept binary as well.
The file_name also needs to be passed as separate argument