Refactor oscilloscope.py
Createc instance will give error if data_producer is in a child thread.
Now it is in the main thread with the bokeh grapher.

Createc instance will now only be initiated once for the data_producer,
unlike before where it is initiated every 0.2 second.

data_producer and DT670 are now sub-modules of utils.

The docstrings for the above two are also updated.


Former-commit-id: 8a9208f481fcd2037bb18d8ce6d4f3179ab2baca